+++ types/src/errors.d.ts
@@ -64,6 +64,14 @@ declare class InvalidCallbackError extends TypeError {
      */
     constructor(where: string, value: unknown);
 }
+declare class ReadonlySignalError extends Error {
+    /**
+     * Constructs a new ReadonlySignalError.
+     *
+     * @param where - The location where the error occurred.
+     */
+    constructor(where: string);
+}
 /**
  * Error thrown when an API requiring an owner is called without one.
  */
@@ -82,4 +90,4 @@ declare function validateSignalValue<T extends {}>(where: string, value: unknown
 declare function validateReadValue<T extends {}>(where: string, value: T | null | undefined): asserts value is T;
 declare function validateCallback(where: string, value: unknown): asserts value is (...args: unknown[]) => unknown;
 declare function validateCallback<T>(where: string, value: unknown, guard: (value: unknown) => value is T): asserts value is T;
-export { type Guard, CircularDependencyError, NullishSignalValueError, InvalidSignalValueError, UnsetSignalValueError, InvalidCallbackError, RequiredOwnerError, DuplicateKeyError, validateSignalValue, validateReadValue, validateCallback, };
+export { type Guard, CircularDependencyError, NullishSignalValueError, InvalidSignalValueError, UnsetSignalValueError, InvalidCallbackError, ReadonlySignalError, RequiredOwnerError, DuplicateKeyError, validateSignalValue, validateReadValue, validateCallback, };